Setup of Sorting
All main pieces are in place as a draft. Objects can now be dragged but no dropped into place yet. When items are released from drag they return to their position. Items were meant to swap places if landed ontop one another but still not working for some reason, gonna figure that out later. All sprites are just examples (they are not going to be in our publish game).