Sound, TransformingSprites, etc
The sprites that change arent the proper sprites yet, they are just examples. Sound has been added when item is picked up and dropped. Still working on to make items disappear once in washer and bin, right now they still stay on top. Also need to work on when all items are in, soap animation must be dragged and animated to let the washer start washing the clothes and then puzzle = complete.